Add files via upload
Hello Everyone, 

This post is regarding about the #codebasicsresumeprojectchallenge. In this challenge they gave us a dataset of an Hardware Company named Atilq_Hardware. In this data set they provided us with 6 Tables as below: 
•	dim_customer: contains customer-related data
•	dim_product: contains product-related data
•	fact_gross_price: contains gross price information for each product
•	fact_manufacturing_cost: contains the cost incurred in the production of each product
•	fact_pre_invoice_deductions: contains pre-invoice deductions information for each product
•	fact_sales_monthly: contains monthly sales data for each product.
The Domain:  Consumer Goods and Function: Executive Management
Also provided us with a set of 10 questions ranging from easy to hard and asked us to provide us with the solutions and present it in an understandable way.